refactor: 适配新版 telegram-bot-api 并优化交互

+++ src/plugins/gatekeeper/request_verify_handle.go
@@ -112,7 +112,7 @@ func VerifyRequestMember(update tgbotapi.Update) {
 	go requestVerify(chatId, userId, photo.MessageID, name)
 }
 
-func requestVerify(chatId int64, userId int64, messageId int, name string) {
+func requestVerify(chatId int64, userId int64, messageId int64, name string) {
 	time.Sleep(time.Minute)
 	if has, _ := verifySet.checkExistAndRemove(userId, chatId); !has {
 		return
